# feat: implement optimised insert_range