--- a/Cargo.toml
+++ b/Cargo.toml
@@ -22,6 +22,8 @@ byteorder = "1.0"
 
 [dev-dependencies]
 criterion = "0.3"
+quickcheck = "0.9"
+quickcheck_macros = "0.9"
 
 [[bench]]
 name = "lib"