--- a/benches/lib.rs
+++ b/benches/lib.rs
@@ -206,6 +206,30 @@ fn remove_range_bitmap(c: &mut Criterion) {
     });
 }
 
+fn insert_range_bitmap(c: &mut Criterion) {
+    for &size in &[10, 100, 1_000, 5_000, 10_000, 20_000] {
+        let mut group = c.benchmark_group("insert_range");
+        group.throughput(criterion::Throughput::Elements(size));
+        group.bench_function(format!("from_empty_{}", size), |b| {
+            let bm = RoaringBitmap::new();
+            b.iter_batched(
+                || bm.clone(),
+                |mut bm| black_box(bm.insert_range(0..size)),
+                criterion::BatchSize::SmallInput,
+            )
+        });
+        group.bench_function(format!("pre_populated_{}", size), |b| {
+            let mut bm = RoaringBitmap::new();
+            bm.insert_range(0..size);
+            b.iter_batched(
+                || bm.clone(),
+                |mut bm| black_box(bm.insert_range(0..size)),
+                criterion::BatchSize::SmallInput,
+            )
+        });
+    }
+}
+
 fn iter(c: &mut Criterion) {
     c.bench_function("iter", |b| {
         let bitmap: RoaringBitmap = (1..10_000).collect();
@@ -300,6 +324,7 @@ criterion_group!(
     is_subset,
     remove,
     remove_range_bitmap,
+    insert_range_bitmap,
     iter,
     is_empty,
     serialize,