--- a/src/bitmap/container.rs
+++ b/src/bitmap/container.rs
@@ -41,10 +41,8 @@ impl Container {
     pub fn insert_range(&mut self, range: Range<u16>) -> u64 {
         // If the range is larger than the array limit, skip populating the
         // array to then have to convert it to a bitmap anyway.
-        if matches!(self.store, Store::Array(_)) {
-            if range.end - range.start > ARRAY_LIMIT as u16 {
-                self.store = self.store.to_bitmap()
-            }
+        if matches!(self.store, Store::Array(_)) && range.end - range.start > ARRAY_LIMIT as u16 {
+            self.store = self.store.to_bitmap()
         }
 
         let inserted = self.store.insert_range(range);