--- a/src/bitmap/inherent.rs
+++ b/src/bitmap/inherent.rs
@@ -43,6 +43,81 @@ impl RoaringBitmap {
         container.insert(index)
     }
 
+    pub fn insert_range(&mut self, range: Range<u64>) -> u64 {
+        assert!(
+            range.end <= u64::from(u32::max_value()) + 1,
+            "can't index past 2**32"
+        );
+        if range.is_empty() {
+            return 0;
+        }
+
+        let (start_container_key, start_index) = util::split(range.start as u32);
+        let (end_container_key, end_index) = util::split((range.end) as u32);
+
+        // Find the container index for start_container_key
+        let start_i = match self
+            .containers
+            .binary_search_by_key(&start_container_key, |c| c.key)
+        {
+            Ok(loc) => loc,
+            Err(loc) => {
+                self.containers
+                    .insert(loc, Container::new(start_container_key));
+                loc
+            }
+        };
+
+        // If the end range value is in the same container, just call into
+        // the one container.
+        if start_container_key == end_container_key {
+            return self.containers[start_i].insert_range(start_index..end_index);
+        }
+
+        // For the first container, insert start_index..u16::MAX, with
+        // subsequent containers inserting 0..MAX.
+        //
+        // The last container (end_container_key) is handled explicitly outside
+        // the loop.
+        let mut low = start_index;
+        let mut inserted = 0;
+
+        // Walk through the containers until the container for end_container_key
+        let end_i = usize::from(end_container_key - start_container_key);
+        for i in start_i..end_i {
+            // Fetch (or upsert) the container for i
+            let c = match self.containers.get_mut(i) {
+                Some(c) => c,
+                None => {
+                    // For each i, the container key is start_container + i in
+                    // the upper u8 of the u16.
+                    let key = start_container_key + ((1 << 8) * i) as u16;
+                    self.containers.insert(i, Container::new(key));
+                    &mut self.containers[i]
+                }
+            };
+
+            // Insert the range subset for this container
+            inserted += c.insert_range(low..u16::MAX);
+
+            // After the first container, always fill the containers.
+            low = 0;
+        }
+
+        // Handle the last container
+        let c = match self.containers.get_mut(end_i) {
+            Some(c) => c,
+            None => {
+                let (key, _) = util::split(range.start as u32);
+                self.containers.insert(end_i, Container::new(key));
+                &mut self.containers[end_i]
+            }
+        };
+        c.insert_range(0..end_index);
+
+        inserted
+    }
+
     /// Adds a value to the set.
     /// The value **must** be greater or equal to the maximum value in the set.
     ///
@@ -131,7 +206,7 @@ impl RoaringBitmap {
             range.end <= u64::from(u32::max_value()) + 1,
             "can't index past 2**32"
         );
-        if range.start == range.end {
+        if range.is_empty() {
             return 0;
         }
         // inclusive bounds for start and end
@@ -290,5 +365,63 @@ impl RoaringBitmap {
 impl Default for RoaringBitmap {
     fn default() -> RoaringBitmap {
         RoaringBitmap::new()
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use quickcheck_macros::quickcheck;
+
+    #[quickcheck]
+    fn insert_range(r: Range<u32>, checks: Vec<u32>) {
+        let r: Range<u64> = u64::from(r.start)..u64::from(r.end);
+
+        let mut b = RoaringBitmap::new();
+        let inserted = b.insert_range(r.clone());
+        if r.end > r.start {
+            assert_eq!(inserted, r.end - r.start);
+        } else {
+            assert_eq!(inserted, 0);
+        }
+
+        // Assert all values in the range are present
+        for i in r.clone() {
+            assert!(b.contains(i as u32), format!("does not contain {}", i));
+        }
+
+        // Run the check values looking for any false positives
+        for i in checks {
+            let bitmap_has = b.contains(i);
+            let range_has = r.contains(&u64::from(i));
+            assert!(
+                bitmap_has == range_has,
+                format!(
+                    "value {} in bitmap={} and range={}",
+                    i, bitmap_has, range_has
+                )
+            );
+        }
+    }
+
+    #[test]
+    fn test_insert_range_same_container() {
+        let mut b = RoaringBitmap::new();
+        let inserted = b.insert_range(1..5);
+        assert_eq!(inserted, 4);
+
+        for i in 1..5 {
+            assert!(b.contains(i));
+        }
+    }
+
+    #[test]
+    fn test_insert_range_pre_populated() {
+        let mut b = RoaringBitmap::new();
+        let inserted = b.insert_range(1..20_000);
+        assert_eq!(inserted, 19_999);
+
+        let inserted = b.insert_range(1..20_000);
+        assert_eq!(inserted, 0);
     }
 }