@Kerollmops 2020-12-03
There seems the CI is not happy, more specifically clippy. Could you apply the requested changes, and do a local `cargo clippy` too, please? :)

https://github.com/RoaringBitmap/roaring-rs/actions/runs/398349871