@Kerollmops 2020-12-04
Hum it is seems that clippy wants you to use collect instead of `from_iter`, sorry for that but could you change it, please? :)

https://github.com/RoaringBitmap/roaring-rs/runs/1497760709?check_suite_focus=true#step:7:62