@domodwyer 2020-12-04
I don't use `from_iter` anywhere - looks like the CI has been failing since 9th November: https://github.com/RoaringBitmap/roaring-rs/actions/runs/354808367