@Kerollmops 2020-12-04
Ok, strange, Clippy seems to have been updated since the last PR I merged. I will do a PR to fix that issue. Sorry about that.