@Kerollmops 2020-12-04
Not sure I found the best way to fix those Clippy recommendations, I just let the Clippy warnings pass instead of using `-D warnings`.

https://github.com/RoaringBitmap/roaring-rs/pull/77