@domodwyer 2020-12-05
Ah these lints are legit but were my intent! I'll stick an `allow` on them 👍