@domodwyer 2020-12-05
I don't get these lint failures locally, so hopefully the changes will allow them to pass - bit of a blind change - thanks for sticking with me to get this merged!