@Kerollmops 2020-12-05
Ok, so [the CI is complaining again](https://github.com/RoaringBitmap/roaring-rs/actions/runs/402518372) and it seems that it is related to the work you have done in #78, I am sorry but it seems that bors rebase but **after** testing the branch.

Could you rebase your changes, I am sorry that this PR is so long to merge.