@Kerollmops 2020-12-06
Thank you for your work and your perseverance. I will bump and release a new roaring version 🎉